workhorse: initial add of Sidekiq UI assets
Relates to charts/gitlab#689

This adds the assets needed for the Sidekiq UI on `/admin/background_jobs`
page of GitLab UI (falls under `/admin/sidekiq`).

This is currently not ideal, because if the ruby version, or Sidekiq
gem version are changed, this Dockerfile needs updated. We'll need to either
detect the versins, or use a properly crafter `COPY` that globs the paths
safely. Such a glob, if possible, should only accept differences in the
numeric identifiers