workhorse: pull sidekiq assets from gitlab-rails
To address charts/gitlab#689

Strip down `/srv/gitlab/vendor/bundle/ruby/*` to _just_ the assets from
Sidekiq gem, and copy thos to the matching location in Workhorse.

This is ugly, but the cleanest method I could fine of pulling the needed assets
in without modifying external code, which is less clean than this modification.