task-runner: rework backup exists checks
Rework the checks in `backup` function to check for directory being
empty, as opposed to existence.

While s3cmd will create the target directory if not present, gsutil
requires that the target exists. To this end, pre-create the target
diretory, and then check if there is anything in it after (r)sync.

Signed-off-by: Jason Plum <jplum@gitlab.com>